+++ homeassistant/components/gentex_homelink/const.py
@@ -3,3 +3,6 @@
 DOMAIN = "gentex_homelink"
 OAUTH2_TOKEN = "https://auth.homelinkcloud.com/oauth2/token"
 POLLING_INTERVAL = 5
+
+EVENT_PRESSED = "Pressed"
+EVENT_OFF = "Off"
