+++ homeassistant/components/gentex_homelink/coordinator.py
@@ -11,11 +11,10 @@ from homelink.provider import Provider
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 import homeassistant.helpers.device_registry as dr
-from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
-from .binary_sensor import HomeLinkBinarySensor
-from .const import DOMAIN, POLLING_INTERVAL
+from .const import POLLING_INTERVAL
+from .event import HomeLinkEventEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -52,7 +51,7 @@ class HomeLinkCoordinator(DataUpdateCoordinator[dict[str, Any]]):
         self.last_sync_timestamp = None
         self.last_sync_id = None
         self.config_entry = config_entry
-        self.buttons = []
+        self.buttons: list[HomeLinkEventEntity] = []
 
     async def _async_setup(self):
         await self.provider.enable()
@@ -60,16 +59,8 @@ class HomeLinkCoordinator(DataUpdateCoordinator[dict[str, Any]]):
         device_data = await self.provider.discover()
 
         for device in device_data:
-            device_info = DeviceInfo(
-                identifiers={
-                    # Serial numbers are unique identifiers within a specific domain
-                    (DOMAIN, device.id)
-                },
-                name=device.name,
-            )
-
             self.buttons = [
-                HomeLinkBinarySensor(b.id, b.name, device_info, self)
+                HomeLinkEventEntity(b.id, b.name, device.id, device.name, self)
                 for b in device.buttons
             ]
